[FIX] translation warning: int to size_t

+++ src/utils/translate.cpp
@@ -1,7 +1,7 @@
 #include "translate.h"
 
 std::vector<GLfloat> translate(std::vector<GLfloat> object, float x, float y, float z) {
-  for (int i = 0; i < object.size(); i++)
+  for (size_t i = 0; i < object.size(); i++)
   {
     object[i] += x;
     object[i + 1] += y;
